refactor: extract test blocks to sibling files
Behavior-preserving relocation per the sibling-tests convention
(`#[cfg(test)] mod tests;` + flat `use super::*;`), shrinking the three
largest in-file test blocks:

- layout/mod.rs    839 -> 194 lines (646-line test block -> layout/tests.rs)
- pose.rs -> pose/ (mod.rs production + tests.rs; 608-line test block)
- pixel_painter/mod.rs 2736 -> 1709 lines (-37%; -> pixel_painter/tests.rs)

Each moved block keeps its original `use super::*;`, so it is semantically
identical to the inline module. Docs updated (workspace tree diagram,
pixtuoid-core + tui CLAUDE.md) per docs-currency. 823 tests pass.
